+++ src/common/utils/__tests__/Utils.test.ts
@@ -91,7 +91,6 @@ describe('Utils', () => {
       { id: '1', name: 'item 1' },
       { id: '2', name: 'item 2' },
       { id: '3', name: 'item 3' },
-      { id: '4' },
     ];
 
     it('should return true when there is a matching value at the key', () => {