+++ src/player/components/Timeline.vue
@@ -31,7 +31,8 @@
       v-for="timestamp of activeTimestamps"
       :key="`t${timestamp.id}`"
       class="Timestamp"
-      :src="isEditing ? 'ic_timestamp_draft.svg' : 'ic_timestamp.svg'"
+      :class="timestampClass(timestamp)"
+      :src="timestampIcon(timestamp)"
       :style="{ left: `${(timestamp.at / duration) * 100}%` }"
     />
     <VueSlider
@@ -154,6 +155,18 @@ export default class Timeline extends Mixins(VideoControllerMixin, KeyboardShort
     }
 
     this.goToNextTimestampOnTimeChange(newTime);
+  }
+
+  public timestampClass(timestamp: Api.AmbigousTimestamp): Record<string, boolean> {
+    return {
+      active: timestamp.id === this.activeTimestamp?.id,
+    };
+  }
+
+  public timestampIcon(timestamp: Api.AmbigousTimestamp): string {
+    if (!this.isEditing) return 'ic_timestamp.svg';
+    if (!timestamp.edited) return 'ic_timestamp_draft.svg';
+    return 'ic_timestamp_draft_edited.svg';
   }
 
   /**
@@ -280,6 +293,10 @@ $translationInactiveSliderVrv: 3px;
     width: 12px;
     top: 4px;
     transform: translateX(-50%);
+
+    &.active {
+      transform: translateX(-50%) translateY(-12px);
+    }
   }
 }